ARCH-1: extract views dispatch arm into commands::views::run (iter-225)
